Major journalling re-work, breaks existing API
The journal spec wasn't flexible enough to allow for multiple operations
in the same scope, but on different instances of that scope. IE, it was
not possible to operate on the transactions collection of two different
accounts (the scope) in the same journal.

This patchset modified the journal spec to greatly simplify the parsing
and playing-back of journals by using a simple format that represents an array
of records. Each record maps a collection to a scope instance, and can
contain any number of collection operations (create, update, delete).